Use int instead of uint8_t to represent bool

+++ src/binaryen-c.cpp
@@ -2134,7 +2134,7 @@ BinaryenExpressionRef BinaryenSwitchGetValue(BinaryenExpressionRef expr) {
   return static_cast<Switch*>(expression)->value;
 }
 // Call
-uint8_t BinaryenCallIsReturn(BinaryenExpressionRef expr) {
+int BinaryenCallIsReturn(BinaryenExpressionRef expr) {
   if (tracing) {
     std::cout << "  BinaryenCallIsReturn(expressions[" << expressions[expr]
               << "]);\n";
@@ -2177,7 +2177,7 @@ BinaryenExpressionRef BinaryenCallGetOperand(BinaryenExpressionRef expr,
   return static_cast<Call*>(expression)->operands[index];
 }
 // CallIndirect
-uint8_t BinaryenCallIndirectIsReturn(BinaryenExpressionRef expr) {
+int BinaryenCallIndirectIsReturn(BinaryenExpressionRef expr) {
   if (tracing) {
     std::cout << "  BinaryenCallIndirectIsReturn(expressions["
               << expressions[expr] << "]);\n";
